Retrieve UC4 Application by name instead of id
ARA V4 now exports some of their domain objects using "system_name" instead of
system_id. The earlier Jenkins plugin relied on using id for getting the application
name, This is now modified to use system_name.

Also, the export XMLs now seem to be producing UTF-16 encoded XMLs rather than
UTF-8 as it was in ARA V3.5. This caused parsing errors like "Content not allowed
in prolog" as the content was actually UTF-8 but the encoding in the XML was mentioned
as UTF-16. Reading it as a FileReader made it work with UTF-16 encoded XMLs.